PZTRAC-4 Fixed monetary type

+++ Model/AbstractTransaction.cs
@@ -4,7 +4,7 @@ namespace _2023pz_trrepo.Model
 {
     public abstract class AbstractTransaction
     {
-        protected AbstractTransaction(string title, int amount, long walletId, long iconId, string? description="")
+        protected AbstractTransaction(string title, double amount, long walletId, long iconId, string? description="")
         {
             Title = title;
             Description = description;
@@ -16,7 +16,7 @@ namespace _2023pz_trrepo.Model
         public long Id { get; set; }
         public string Title { get; set; }
         public string? Description { get; set; }
-        public int Amount { get; set; }
+        public double Amount { get; set; }
         public long WalletId { get; set; }
         public long IconId { get; set; }
         public abstract string TransactionType();
